pyproject.toml
[tool.poetry]
name = "llmchatbot"
version = "0.1.0"
description = ""
authors = ["kevin <ftkevin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.2.6"
langchain-openai = "^0.1.14"
fastapi = "^0.111.0"
python-dotenv = "^1.0.1"
uvicorn = "^0.30.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
